feature/dynamic driver



See merge request !6

+++ duneuro/io/volume_conductor_reader.hh
@@ -68,7 +68,7 @@ namespace duneuro
         It endit = gv.template end<0>();
         for (; it != endit; ++it) {
           auto pe = elementIndexToPhysicalEntity[factory.insertionIndex(*it)];
-          if (pe-offset >= tensors.size()) {
+          if (pe-offset >= static_cast<int>(tensors.size())) {
             DUNE_THROW(Dune::Exception, "physical entitiy of element "
                                             << factory.insertionIndex(*it) << " is " << pe
                                             << " but only " << tensors.size()
